added client program
