+++ src/Main.scala
@@ -1,4 +1,4 @@
-import java.util.concurrent.{Executors, PriorityBlockingQueue}
+import java.util.concurrent.{ExecutorService, Executors, PriorityBlockingQueue}
 
 import scala.concurrent.ExecutionContext
 import scala.util.Try
@@ -9,33 +9,26 @@ object Main {
     override def compare(x: ScheduledTask, y: ScheduledTask): Int = x.timestamp compare y.timestamp
   }
 
-  val SECOND_IN_MILLI = 1000
-
   def main(args: Array[String]): Unit = {
 
-    implicit val ec = ExecutionContext.fromExecutorService(Executors.newFixedThreadPool(10))
+    implicit val ec = ExecutionContext.fromExecutorService(Executors.newFixedThreadPool(5))
 
-    val scheduler = new Scheduler {
-      val size: Int = Try(args(0)).map(_.toInt).getOrElse(100)
-      // Java's PriorityQueue is MIN heap by default
-      val minHeap = new PriorityBlockingQueue[ScheduledTask](size, TimeOrder)
+    val size: Int = Try(args(0)).map(_.toInt).getOrElse(100)
+    // Java's PriorityQueue is MIN heap by default
+    val minHeap = new PriorityBlockingQueue[ScheduledTask](size, TimeOrder)
 
-      override def schedule(task: () => Unit, timestamp: Long): Unit = {
-        minHeap.offer(ScheduledTask(task, timestamp))
-      }
+    val scheduler = new SchedulerImp(minHeap)
 
-      override def start()(implicit ec: ExecutionContext): Unit = {
-        while(true) {
-          // put and take are blocking
-          val nextTask = minHeap.take()
-          if (nextTask.timestamp < System.currentTimeMillis) nextTask.task() else minHeap.put(nextTask)
-
-        }
-      }
-    }
+    val SECOND_IN_MILLI = 1000
+    val exec: ExecutorService = Executors.newFixedThreadPool(5)
 
     for (gap <- 0 to SECOND_IN_MILLI * 10 by SECOND_IN_MILLI) {
-      scheduler.schedule(() => println(gap), System.currentTimeMillis() + gap)
+      exec.execute(new Runnable() {
+        override def run(): Unit = {
+          val threadId = Thread.currentThread.getId
+          scheduler.schedule(() => println(s"Thread # = $threadId, Gap = $gap"), System.currentTimeMillis() + gap)
+        }
+      })
     }
 
     scheduler.start()
