+++ src/Scheduler.scala
@@ -1,3 +1,5 @@
+import java.util.concurrent.PriorityBlockingQueue
+
 import scala.concurrent.ExecutionContext
 
 //type Task = () => Unit
@@ -5,4 +7,20 @@ import scala.concurrent.ExecutionContext
 trait Scheduler {
   def schedule(task: () => Unit, timestamp: Long): Unit
   def start()(implicit ec: ExecutionContext): Unit
+}
+
+
+class SchedulerImp(val minHeap: PriorityBlockingQueue[ScheduledTask]) extends Scheduler {
+
+  override def schedule(task: () => Unit, timestamp: Long): Unit = {
+    minHeap.offer(ScheduledTask(task, timestamp))
+  }
+
+  override def start()(implicit ec: ExecutionContext): Unit = {
+    while(true) {
+      // put and take are blocking
+      val nextTask = minHeap.take()
+      if (nextTask.timestamp < System.currentTimeMillis) nextTask.task() else minHeap.put(nextTask)
+    }
+  }
 }
